3-11-16
Changed string formatting to align data being printed.
Need to possibly remove capitalised name from database as it is
duplicated. Possibly capitalise in program.  Or remove the lowercase
names and just use the capitalised data from the database, which can be
converted to lowercase...
If not possible just point out why it isn't possible. And talk about
other places where it is efficient.